Script update check is now more robust
Instead of using the HTML on Github to find the commit hashes the script
now uses the Github API which means that there should now be less fixes
to the script to accommodate the changes in the Github source code.

See the README file for more information about the new api_token
variable.